ai-service/requirements.txt
# FastAPI and server dependencies
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0
python-multipart==0.0.6

# AI/ML dependencies
openai==1.10.0
anthropic==0.8.1
langchain==0.1.4
langchain-openai==0.0.5
langchain-anthropic==0.1.1
pinecone-client==3.0.2

# Utilities
python-dotenv==1.0.0
httpx==0.26.0
tenacity==8.2.3
redis==5.0.1

# Monitoring and logging
structlog==24.1.0
sentry-sdk[fastapi]==1.40.0
